♻️ Move getLogo.ts to assets.ts
