+++ src/app/[locale]/layout.tsx
@@ -6,7 +6,7 @@ import { HeaderContainer } from '@/app/ui/templates/header/container'
 import { LayoutWrapper } from '@/app/ui/templates/layout-wrapper'
 import { EMAIL, LANGUAGE, LOGO_TITLE, PHONE_NUMBER, TWITTER_HANDLE, TWITTER_ID } from '@/constants'
 import { routing } from '@/i18n/routing'
-import { getLogo } from '@/utils/logo'
+import { getLogo } from '@/utils/assets'
 import classNames from 'classnames'
 import type { Metadata } from 'next'
 import { NextIntlClientProvider } from 'next-intl'
